+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ContourMixin: is_adjacent
 - Wire2D: area
 - Circle2D: bsplinecurve_intersections
+- add tolerance param to many methods from edges and wires.
 
 ### Fixed
 - 2D conversion: create 2D function name in core_compiled